requirements.txt
asgiref==3.7.2
Django==4.2.5
django-clearcache==1.2.1
django-shopping-cart==0.1
Pillow==10.0.0
sqlparse==0.4.4
tzdata==2023.3
